Fixed review issues (thanks @hrissan)
* Upstream graceful shutdown now works
* Harmful timeout handlers removed
* Forward upstream connection (handshake) error
